Cleaned up the constructor. Reducing number of lines of code by moving definitions inline in class definition.

+++ AudioTools/AudioLevel.cpp
@@ -28,8 +28,8 @@ void AudioLevel::paintEvent(QPaintEvent *event) {
     QPainter painter(this);
     // draw level
     qreal widthLevel = m_level * width();
-    painter.fillRect(0, 0, widthLevel, height(), Qt::red);
+    painter.fillRect(0, 0, (int) widthLevel, height(), Qt::red);
     // clear the rest of the control
-    painter.fillRect(widthLevel, 0, width(), height(), Qt::black);
+    painter.fillRect((int) widthLevel, 0, width(), height(), Qt::black);
 
 }
